% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{dropout}
\alias{dropout}
\title{Implement Drop-out Mechanism}
\usage{
dropout(dat, type_dropout = NULL, prop.miss = 0.3, stochastic.component = 0.2)
}
\arguments{
\item{type_dropout}{pass character or character vector specifying the type of drop-out you want,
options are c('mcar', 'cmcar', 'mar', 'mnar')}

\item{prop.miss}{proportion missing, either a numeric scalar or vector, if vector, must be equal to number of timepoints
if it's a scalar it will be the proportion at the last timepoint, amount at other timepoints will be equally spread out}

\item{stochastic.component}{control how deterministic drop-out is by adding noise to the governing parameter;
this is just the variance of a standard normal}

\item{data}{pass a dataframe generated using function \code{sim_dat()}}
}
\value{
returns a dataframe containing the simulated data, with the original scores and the new scores with missingness
}
\description{
Implement Drop-out Mechanism
}
\details{
Pass the dataframe of data and implement different types of drop-out to yield missing data
that corresponds to different missing data categories
TODO: Add checks, like if 'cmcar' is selected, and there's no covariate, kick error
cannot implement dropout at baseline currently - should that be an option?
}
