% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_ord_logistic.R
\name{sim_dat_ord_logistic}
\alias{sim_dat_ord_logistic}
\title{Simulate Longitudinal Ordinal Data with Logistic Specification}
\usage{
sim_dat_ord_logistic(
  N = 100,
  number.groups = 2,
  number.timepoints = 4,
  reg.formula = NULL,
  Beta = 0,
  thresholds = c(-2, -1, 0, 1),
  corr = "ar1",
  cor.value = NULL,
  cond.mcar = F,
  Covariate = F
)
}
\arguments{
\item{reg.formula}{this is a regression formula to pass to the data generation; null is
formula(~ Group + Time + Time*Group),}

\item{Beta}{this is the Beta for the regression equation; numeric matrix of Beta values OR a scalar value != 0
that will be the final value of the interaction parameters, default is all(Beta == 0) for Type I error simulations}

\item{thresholds}{the ordinal generation uses a logistic approach,
and these are the tresholds}

\item{corr}{options for correlation structure  c('ind', 'ar1', 'cs'), default is 'ar1'}

\item{cor.value}{numeric, the first corr in ar1, the corr in cs option}

\item{cond.mcar}{logical; do you want a conditional MCAR data generation}

\item{Covariate}{logical; do you want a random simulated covariate?}
}
\value{
returns a dataframe containing the simulated data
}
\description{
TODO: Build in check if the correlations are too high! Where is that code?
Note that the variances aren't used if passed,
Requires MASS R package
}
