% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_types.R
\name{sim_dat_types}
\alias{sim_dat_types}
\title{Simulate Conditional Longitudinal Data
This is a conditional model with a random intercept}
\usage{
sim_dat_types(
  N = 1000,
  data.type = NULL,
  number.groups = 2,
  number.timepoints = 4,
  reg.formula = NULL,
  Beta = NULL,
  sigma2 = NULL,
  shape2 = NULL,
  phi = NULL,
  theta = NULL,
  zip = NULL,
  thresholds = NULL,
  k = NULL,
  subject.var = 1
)
}
\arguments{
\item{data.type}{Specify the type of data you want to generate. Select
one from the following: 'Ordinal', 'Beta', 'Gaussian', 'NegBinom',
'ZIP', 'ZINB', 'Multinom'. This must be a character vector of length 1.}

\item{reg.formula}{this is a regression formula to pass to the data generation; null is
formula(~ Group + Time + Time*Group),}

\item{Beta}{this is the Beta for the regression equation; numeric matrix of Beta values OR a scalar value != 0
that will be the final value of the interaction parameters, default is all(Beta == 0) for Type I error simulations}

\item{sigma2}{the residual error in Gaussian data}

\item{shape2}{parameter in Beta distribution}

\item{phi}{parameter in Beta distribution}

\item{theta}{size parameter in NB}

\item{zip}{zero inflation parameter}

\item{thresholds}{thresholds in ordinal data}

\item{k}{number of categories in ordinal/nominal}

\item{subject.var}{the subject-level variance, also known as the variance
associated with the random intercept}
}
\value{
returns a dataframe containing the simulated data
}
\description{
Select the data type you want to generate - this is best suited for
fitting data in \code{glmmTMB()} among other R packages that are based on
the conditional model with a random intercept
}
