% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCov_gls.R
\name{VarCov_gls}
\alias{VarCov_gls}
\title{Extract Variance-Covariance Matrix from nlme object}
\usage{
VarCov_gls(fit)
}
\arguments{
\item{fit}{nlme model, specifically a MMRM model estimated using \code{gls()} function}
}
\value{
returns a matrix of the estimated variance covariance values
}
\description{
Requires nlme object
}
